added back changes from previous pr and started writing tests in document upload files
