+++ app/sidekiq/lighthouse/benefits_documents/document_upload.rb
@@ -4,9 +4,8 @@ require 'ddtrace'
 require 'timeout'
 require 'lighthouse/benefits_documents/worker_service'
 require 'lighthouse/benefits_documents/constants'
-require 'lighthouse/failure_notification'
 
-class Lighthouse::DocumentUpload
+class Lighthouse::BenefitsDocuments::DocumentUpload
   include Sidekiq::Job
   extend SentryLogging
 
@@ -14,11 +13,6 @@ class Lighthouse::DocumentUpload
 
   FILENAME_EXTENSION_MATCHER = /\.\w*$/
   OBFUSCATED_CHARACTER_MATCHER = /[a-zA-Z\d]/
-
-  DD_ZSF_TAGS = ['service:claim-status', 'function: evidence upload to Lighthouse'].freeze
-
-  NOTIFY_SETTINGS = Settings.vanotify.services.benefits_management_tools
-  MAILER_TEMPLATE_ID = NOTIFY_SETTINGS.template_id.evidence_submission_failure_email
 
   # retry for  2d 1h 47m 12s
   # https://github.com/sidekiq/sidekiq/wiki/Error-Handling
@@ -29,27 +23,28 @@ class Lighthouse::DocumentUpload
   end
 
   sidekiq_retries_exhausted do |msg, _ex|
-    # There should be 2 values in msg['args']:
-    # 1) The ICN of the user
-    # 2) Document metadata
-
-    next unless Flipper.enabled?('cst_send_evidence_failure_emails')
-
-    icn = msg['args'].first
+    job_id = msg['jid']
+    job_class = 'Lighthouse::BenefitsDocuments::DocumentUpload'
     first_name = msg['args'][1]['first_name'].titleize
+    claim_id = msg['args'][1]['claim_id']
+    tracked_item_id = msg['args'][1]['tracked_item_id']
     filename = obscured_filename(msg['args'][1]['file_name'])
     date_submitted = format_issue_instant_for_mailers(msg['created_at'])
     date_failed = format_issue_instant_for_mailers(msg['failed_at'])
+    upload_status = BenefitsDocuments::Constants::UPLOAD_STATUS[:FAILED]
+    uuid = msg['args'][1]['uuid']
+    user_account = UserAccount.find_or_create_by(id: uuid)
+    personalisation = { first_name:, filename:, date_submitted:, date_failed: }
 
-    Lighthouse::FailureNotification.perform_async(icn, first_name, filename, date_submitted, date_failed)
-
-    ::Rails.logger.info('Lighthouse::DocumentUpload exhaustion handler email queued')
-    StatsD.increment('silent_failure_avoided_no_confirmation', tags: DD_ZSF_TAGS)
-  rescue => e
-    ::Rails.logger.error('Lighthouse::DocumentUpload exhaustion handler email error',
-                         { message: e.message })
-    StatsD.increment('silent_failure', tags: DD_ZSF_TAGS)
-    log_exception_to_sentry(e)
+    EvidenceSubmission.create(
+      job_id:,
+      job_class:,
+      claim_id:,
+      tracked_item_id:,
+      upload_status:,
+      user_account:,
+      template_metadata_ciphertext: { personalisation: }.to_json
+    )
   end
 
   def self.obscured_filename(original_filename)
@@ -73,10 +68,6 @@ class Lighthouse::DocumentUpload
 
     # We display dates in mailers in the format "May 1, 2024 3:01 p.m. EDT"
     timestamp.strftime('%B %-d, %Y %-l:%M %P %Z').sub(/([ap])m/, '\1.m.')
-  end
-
-  def self.notify_client
-    VaNotify::Service.new(NOTIFY_SETTINGS.api_key)
   end
 
   def perform(user_icn, document_hash, user_account_uuid, claim_id, tracked_item_id)
