+++ db/migrate/20241217210622_create_evidence_submissions.rb
@@ -1,4 +1,4 @@
-class CreateEvidenceSubmissions < ActiveRecord::Migration[7.1]
+class CreateEvidenceSubmissions < ActiveRecord::Migration[7.2]
   def change
     create_table :evidence_submissions do |t|
       t.string :job_id
@@ -12,6 +12,7 @@ class CreateEvidenceSubmissions < ActiveRecord::Migration[7.1]
       t.string :va_notify_id
       t.string :va_notify_status
       t.date :delete_date
+      t.datetime :va_notify_date
       t.string :tracked_item_id
 
       t.timestamps
