+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2024_12_13_173113) do
+ActiveRecord::Schema[7.2].define(version: 2024_12_17_210622) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "btree_gin"
   enable_extension "fuzzystrmatch"
@@ -645,16 +645,18 @@ ActiveRecord::Schema[7.2].define(version: 2024_12_13_173113) do
     t.string "job_class"
     t.string "request_id"
     t.string "claim_id"
-    t.string "user_account_id"
+    t.uuid "user_account_id", null: false
     t.json "template_metadata_ciphertext"
     t.text "encrypted_kms_key"
     t.string "upload_status"
     t.string "va_notify_id"
     t.string "va_notify_status"
     t.date "delete_date"
+    t.datetime "va_notify_date"
     t.string "tracked_item_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["user_account_id"], name: "index_evidence_submissions_on_user_account_id"
   end
 
   create_table "evss_claims", id: :serial, force: :cascade do |t|
@@ -1740,6 +1742,7 @@ ActiveRecord::Schema[7.2].define(version: 2024_12_13_173113) do
   add_foreign_key "deprecated_user_accounts", "user_accounts"
   add_foreign_key "deprecated_user_accounts", "user_verifications"
   add_foreign_key "education_stem_automated_decisions", "user_accounts"
+  add_foreign_key "evidence_submissions", "user_accounts"
   add_foreign_key "evss_claims", "user_accounts"
   add_foreign_key "form526_submission_remediations", "form526_submissions"
   add_foreign_key "form526_submissions", "user_accounts"
