+++ spec/sidekiq/evss/document_upload_spec.rb
@@ -1,28 +1,53 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+require 'sidekiq/testing'
+Sidekiq::Testing.fake!
 
 require 'evss/document_upload'
 require 'va_notify/service'
 
 RSpec.describe EVSS::DocumentUpload, type: :job do
-  subject { described_class }
+  subject(:job) do
+    described_class.perform_async(auth_headers, user.uuid, document_data.to_serializable_hash)
+  end
+  # subject { described_class }
 
   let(:client_stub) { instance_double('EVSS::DocumentsService') }
+  let(:job_id) { job }
+  let(:job_class) { 'EVSS::DocumentUpload' }
+  let(:claim_id) { '4567' }
+  let(:tracked_item_id) { '1234' }
+  let(:document_type) { 'L023' }
+
+  let(:formatted_submit_date) do
+    # We want to return all times in EDT
+    timestamp = Time.at(issue_instant).in_time_zone('America/New_York')
+
+    # We display dates in mailers in the format "May 1, 2024 3:01 p.m. EDT"
+    timestamp.strftime('%B %-d, %Y %-l:%M %P %Z').sub(/([ap])m/, '\1.m.')
+  end
+  let(:created_at) { DateTime.new(2023, 4, 2) }
+
   let(:notify_client_stub) { instance_double(VaNotify::Service) }
   let(:uploader_stub) { instance_double('EVSSClaimDocumentUploader') }
-
   let(:user_account) { create(:user_account) }
   let(:user_account_uuid) { user_account.id }
   let(:user) { FactoryBot.create(:user, :loa3) }
   let(:filename) { 'doctors-note.pdf' }
   let(:document_data) do
     EVSSClaimDocument.new(
-      evss_claim_id: 189_625,
+      evss_claim_id: claim_id,
       file_name: filename,
-      tracked_item_id: 33,
-      document_type: 'L023'
+      tracked_item_id:,
+      document_type:
     )
+  end
+  let(:personalisation) do
+    { first_name: user.first_name,
+      filename:,
+      date_submitted: created_at.strftime('%B %d, %Y'),
+      date_failed: created_at.strftime('%B %d, %Y') }
   end
   let(:auth_headers) { EVSS::AuthHeaders.new(user).to_h }
 
@@ -34,11 +59,9 @@ RSpec.describe EVSS::DocumentUpload, type: :job do
       'failed_at' => issue_instant
     }
   end
-  let(:tags) { subject::DD_ZSF_TAGS }
 
   before do
     allow(Rails.logger).to receive(:info)
-    allow(StatsD).to receive(:increment)
   end
 
   it 'retrieves the file and uploads to EVSS' do
@@ -52,46 +75,20 @@ RSpec.describe EVSS::DocumentUpload, type: :job do
     described_class.new.perform(auth_headers, user.uuid, document_data.to_serializable_hash)
   end
 
-  context 'when cst_send_evidence_failure_emails is enabled' do
-    before do
-      Flipper.enable(:cst_send_evidence_failure_emails)
-    end
-
-    let(:formatted_submit_date) do
-      # We want to return all times in EDT
-      timestamp = Time.at(issue_instant).in_time_zone('America/New_York')
-
-      # We display dates in mailers in the format "May 1, 2024 3:01 p.m. EDT"
-      timestamp.strftime('%B %-d, %Y %-l:%M %P %Z').sub(/([ap])m/, '\1.m.')
-    end
-
-    it 'calls EVSS::FailureNotification' do
-      subject.within_sidekiq_retries_exhausted_block(args) do
-        expect(EVSS::FailureNotification).to receive(:perform_async).with(
-          user_account.icn,
-          'Bob', # first_name
-          'docXXXX-XXte.pdf', # filename
-          formatted_submit_date, # date_submitted
-          formatted_submit_date # date_failed
-        )
-
-        expect(Rails.logger)
-          .to receive(:info)
-          .with('EVSS::DocumentUpload exhaustion handler email queued')
-        expect(StatsD).to receive(:increment).with('silent_failure_avoided_no_confirmation', tags:)
-      end
-    end
-  end
-
-  context 'when cst_send_evidence_failure_emails is disabled' do
-    before do
-      Flipper.disable(:cst_send_evidence_failure_emails)
-    end
-
-    it 'does not call Lighthouse::Failure Notification' do
-      subject.within_sidekiq_retries_exhausted_block(args) do
-        expect(EVSS::FailureNotification).not_to receive(:perform_async)
-      end
+  it 'creates a failed evidence submission record' do
+    subject.within_sidekiq_retries_exhausted_block(args) do
+      byebug
+      expect(EvidenceSubmission.va_notify_email_not_sent.length).to equal(1)
+      # expect do
+      #   post :create,
+      #        params: { job_id:,
+      #                  job_class:,
+      #                  claim_id:,
+      #                  tracked_item_id:,
+      #                  upload_status: BenefitsDocuments::Constants::UPLOAD_STATUS[:FAILED],
+      #                  user_account:,
+      #                  template_metadata_ciphertext: { personalisation: }.to_json }
+      # end.to change(EvidenceSubmission.va_notify_email_not_sent, :count).by(1)
     end
   end
 end
